Refactor : Option, OptionGroup, Product 일부 수정
- Option, OptionGroup => @Builder 어노테이션 추가, 패키지 경로 일부 변경(model >> persist >> entity)
- Product => OptionGroup 변경 경로 import 반영
