+++ src/main/java/team/mosk/api/server/domain/option/model/persist/OptionGroup.java
@@ -1,9 +1,6 @@
-package team.mosk.api.server.domain.option.model;
+package team.mosk.api.server.domain.option.model.persist;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import team.mosk.api.server.domain.product.model.persist.Product;
 import team.mosk.api.server.global.common.BaseEntity;
 
@@ -14,6 +11,7 @@ import java.util.List;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
 public class OptionGroup extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -28,4 +26,12 @@ public class OptionGroup extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
     private Product product;
+
+    /**
+     * methods
+     */
+
+    public void initProduct(final Product product) {
+        this.product = product;
+    }
 }
