+++ src/main/java/team/mosk/api/server/domain/product/model/persist/Product.java
@@ -2,7 +2,7 @@ package team.mosk.api.server.domain.product.model.persist;
 
 import lombok.*;
 import team.mosk.api.server.domain.category.model.persist.Category;
-import team.mosk.api.server.domain.option.model.OptionGroup;
+import team.mosk.api.server.domain.option.model.persist.OptionGroup;
 import team.mosk.api.server.domain.product.dto.UpdateProductRequest;
 import team.mosk.api.server.domain.product.model.vo.Selling;
 import team.mosk.api.server.domain.store.model.persist.Store;
